feat: Rename 'defauts.js' to 'defaults.js' and remove unnecessary default prompts
